Changed lane to be "put in" the boat All changes: - Moved Boat and BoatType to have own package "boats" - New Boats: PlayerBoat and ComputerBoat for different control methods - Moved Hitbox.java into Tools package - Created Lane.java to hold obstacles and have a hitbox.

+++ core/src/com/dragonboatrace/entities/Entity.java
@@ -31,7 +31,7 @@ public abstract class Entity{
         resize.dispose();
 
         /* Make a new hitbox at the entities position with its width and height */
-        this.box = new Hitbox(pos.x, pos.y, type.getWidth(), type.getHeight());
+        this.box = new Hitbox((int)pos.x, (int)pos.y, type.getWidth(), type.getHeight());
     }
 
     public abstract void update(float deltaTime);
